fetch: moved api -> receipt module
