+++ core/api/src/main/java/com/wansenai/api/receipt/SalesController.java
@@ -10,7 +10,7 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.wansenai.api;
+package com.wansenai.api.receipt;
 
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.wansenai.dto.receipt.sale.QuerySaleOrderDTO;
